#14 classごとにファイルを分け、ボタンを押されたら設定画面が出るように実装
